requirements.txt
streamlit==1.40.1
anthropic
openai
google-generativeai
mistralai
python-dotenv
pyyaml
pandas
jinja2
matplotlib>=3.5.0  # For visualization plots

# Optional: Improved sentence segmentation for claim extraction (v2.0+)
# If not installed, falls back to regex-based splitting
# No model download required - uses blank("en") + sentencizer
spacy>=3.7

# Glass Box Audit Pipeline dependencies
transformers>=4.30.0
torch>=2.0.0
tqdm
tenacity>=8.0.0

# Semantic pre-filtering (Phase 1 optimization)
sentence-transformers>=2.2.2
